Move MvnUtils to fattest.simplicity
